Using generic K dimensional point into the package.
When building a KdTree set the number of dimensions to use for ordering should be provided if the default of 2 is not desired. All the entered point then need to provide that required minimum of coordinates.
Legacy Point2D is left for use when the use of 2 dimensions wants to be enforced.